requirements.txt
Cython==0.29.15
matplotlib==3.3.0
numpy==1.21.6
pandas==1.3.5
pathlib2==2.3.5
pysam==0.19.1
seaborn==0.10.0
pyyaml==6.0